# laravel-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Puntos de Fidelización",
        "description": "API para gestionar los puntos de fidelización de una tarjeta de cliente.",
        "contact": {
            "name": "Ian lopez Zamora",
            "email": "gm.2012.ian.lopez@gmail.com"
        },
        "license": {
            "name": "",
            "url": ""
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/consultar/puntos-periodo-farmacia": {
            "get": {
                "tags": [
                    "Consultar"
                ],
                "summary": "Puntos totales otorgados por la farmacia en un periodo de tiempo",
                "operationId": "826f09697765f0792200fc98975cf4ec",
                "parameters": [
                    {
                        "name": "idFarmacia",
                        "in": "query",
                        "description": "ID Farmacia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dateFrom",
                        "in": "query",
                        "description": "Fecha inicial en formato YYYY-MM-DD HH:MM:SS",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "dateTo",
                        "in": "query",
                        "description": "Fecha final en formato YYYY-MM-DD HH:MM:SS",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saldo consultado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "Id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Nombre": {
                                            "type": "string",
                                            "example": "Farmacia A"
                                        },
                                        "Puntos totales otorgados": {
                                            "type": "integer"
                                        },
                                        "Puntos sin canjear": {
                                            "type": "integer"
                                        },
                                        "Fecha inicial": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "Fecha fin": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "Mensaje": {
                                            "properties": {
                                                "idFarmacia": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "The id farmacia field must be an integer o the id farmacia field is required",
                                                        "type": "string",
                                                        "example": "the id farmacia field is required"
                                                    }
                                                },
                                                "dateFrom": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "The date from field must match the format Y-m-d H:i:s. o the id cliente field is required",
                                                        "type": "string",
                                                        "example": "The date from field must match the format Y-m-d H:i:s."
                                                    }
                                                },
                                                "dateTo": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "The date to field must match the format Y-m-d H:i:s. o the id puntos field is required",
                                                        "type": "string",
                                                        "example": "The date to field must match the format Y-m-d H:i:s."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error de autenticación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "Mensaje": {
                                            "description": "ID de farmacia no existe",
                                            "type": "string",
                                            "example": "El id de farmacia no existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "Mensaje": {
                                            "type": "string",
                                            "example": "Error interno del servidor, contacta con el administrador de la API"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/consultar/saldo-cliente": {
            "get": {
                "tags": [
                    "Consultar"
                ],
                "summary": "Consultar saldo del cliente",
                "operationId": "d3b9a824e4a101752775bd58e63d3a93",
                "parameters": [
                    {
                        "name": "idCliente",
                        "in": "query",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saldo consultado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "Id cliente": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Nombre cliente": {
                                            "type": "string",
                                            "example": "Cliente 1"
                                        },
                                        "Saldo actual": {
                                            "type": "integer",
                                            "example": 12
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error de autenticación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "Mensaje": {
                                            "description": "ID de cliente no existe",
                                            "type": "string",
                                            "example": "El id de cliente no existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos de entrada no válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "Mensaje": {
                                            "properties": {
                                                "idCliente": {
                                                    "type": "array",
                                                    "items": {
                                                        "description": "The id cliente field must be an integer o the id cliente field is required",
                                                        "type": "string",
                                                        "example": "the id cliente field is required"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "Mensaje": {
                                            "type": "string",
                                            "example": "Error interno del servidor, contacta con el administrador de la API"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}